backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cohere==4.32
cryptography==41.0.5
dnspython==2.4.2
fastavro==1.8.2
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
pycparser==2.21
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.7
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
